workshops/requirements.txt
strands-agents
strands-agents-tools
opensearch-py